requirements.txt
pytest
requests
pytest-html
jsonschema
python-dotenv
pytest-xdist
